Refactor: 인수 테스트 코드 beforeEach 중복 제거 리펙터링
