+++ src/test/java/nextstep/subway/AcceptanceTest.java
@@ -1,0 +1,26 @@
+package nextstep.subway;
+
+import io.restassured.RestAssured;
+import nextstep.subway.utils.DatabaseCleanUtil;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
+import org.springframework.boot.web.server.LocalServerPort;
+
+@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
+public class AcceptanceTest {
+
+    @LocalServerPort
+    int port;
+
+    @Autowired
+    DatabaseCleanUtil databaseCleanUtil;
+
+    protected void setUp(){
+        if(RestAssured.port == RestAssured.UNDEFINED_PORT){
+            RestAssured.port = this.port;
+        }
+        databaseCleanUtil.cleanUp();
+    }
+}
